+++ RELEASE-NOTES.md
@@ -1,7 +1,15 @@
+# pyngsi 2.1.3
+## December 22, 2020
+
+- Fixed baseurl not working in SinkOrion
+- Fixed missing asserts in unit tests
+- Added unit tests
+
 # pyngsi 2.1.2
 ## December 22, 2020
 
 - Fixed missing import
+
 # pyngsi 2.1.1
 ## December 21, 2020